refactor: separated out avalil units table into its own component
